Add vocabulary files and building script.
It's a bit hacky, but since we only ever have to use it once, I think it's ok.

Basically the tokenizer already has a vocabulary and indices, but the `Vocabulary` object from
AllenNLP wants to make its own (would be nice to have a `from_dict` function). To work around it,
sort your tokenizer's tokens by their index, feed those into the `Vocabulary` object, and ensure
that the `padding_token` and `oov_token` are appropriately set to `[PAD]` and `[UNK]` in this case.